+++ CHANGELOG.md
@@ -28,6 +28,10 @@ The [history](HISTORY.md) contains the changelogs of older PostBOUND releases.
   The `MostCommonValues` can be used as a drop-in replacement for the old tuple-based API. In addition, it provides more
   high-level methods for working with the most common values.
 - Enabled the MySQL and DuckDB backends to fall back to emulated statistics if the database does not provide them.
+- The `QueryPreparation` API now provides the `projection` and `output` parameters to modify the *SELECT* clause and the type
+  of results to gather for all queries in a more flexible and intuitive way (how did *explain=True*  and *analyze=True*
+  interact?).
+  The old API using *analyze=True*, etc. is now deprecated in favor of these new parameters.
 - The `to_json()` and `to_json_dump()` utilities now support dataclasses out-of-the-box.
 
 ## 🏥 Fixes
@@ -37,7 +41,8 @@ The [history](HISTORY.md) contains the changelogs of older PostBOUND releases.
 - _None_
 
 ## ⚠️ Deprecations
-- _None_
+- The old `QueryPrepration` API using *analyze=True*, etc. is now deprecated in favor of the more flexible *projection* and
+  *output*  parameters. However, we currently have no plans to remove the old API.
 
 ## 🪲 Known bugs
 - The automatic optimization of the Postgres server configuration as part of the Docker installation does not work